Address review findings
- week.ts: new shellWeekStringsNY() derives Sun-anchored topbar +
  Sun–Sat range from weekOfMondayNY(), so the shell follows NY tz
  like the customer side instead of late-Saturday-NY jumping a week
  early on UTC.
- layout.tsx: drop the inline shellWeekStrings helper in favor of
  the lib/week.ts version. Also gate the inventory badge on count > 0
  so "0 listed" doesn't render when the catalog is empty (mirrors the
  orders-badge pattern).
- admin/queries.ts: comment on getOrderingOpen — .single() is
  intentional. Singleton row is invariant; a missing row 500s loudly
  vs silently defaulting.
